+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+* Added possibility to ignore flag to set when using `sink.AddFlagsToSet`, for example:
+
+  ```go
+  sink.AddFlagsToSet(flags, sink.FlagIgnore(sink.FlagFinalBlocksOnly))
+  ```
+
+* Fixed bug when flag are ignored and not defined.
+
 ## v0.2.0
 
 ### Highlights